refactor: move translate fn out
